all courses data scrapped
dfsd

+++ Course_data_scraper/main.py
@@ -2,10 +2,10 @@ import json
 from pathlib import Path
 from json_file_creator import create_courses_json
 
-n = 3
+n = 97
 files_names = []
-courses_dir = 'Physics_courses'
-get_course_requi = False
+courses_dir = 'all_faculties'
+get_course_requi = True
 for i in range(n):
     files_name = f"{courses_dir}/pg{i}.html"
     my_file = Path(files_name)
@@ -18,11 +18,12 @@ if get_course_requi:
     with open(f"{courses_dir}\courses.json", "w", encoding="utf-8") as output_file:
         json.dump(output_data, output_file, ensure_ascii=False, indent=4)
 else:  # to keep existing requirements
-    with open(f"{courses_dir}\courses.json", "r", encoding="utf-8") as output_file:
-        courses_dict = json.load(output_file)
-    for course_num in output_data:
-        for key in output_data[course_num]:
-            if key != 'pre_req' and key != 'parallel_req':
-                courses_dict[course_num][key] = output_data[course_num][key]
-    with open(f"{courses_dir}\courses.json", "w", encoding="utf-8") as output_file:
-        json.dump(courses_dict, output_file, ensure_ascii=False, indent=4)
+    if Path(f"{courses_dir}\courses.json").is_file():
+        with open(f"{courses_dir}\courses.json", "r", encoding="utf-8") as output_file:
+            courses_dict = json.load(output_file)
+        for course_num in output_data:
+            for key in output_data[course_num]:
+                if key != 'pre_req' and key != 'parallel_req':
+                    courses_dict[course_num][key] = output_data[course_num][key]
+        with open(f"{courses_dir}\courses.json", "w", encoding="utf-8") as output_file:
+            json.dump(courses_dict, output_file, ensure_ascii=False, indent=4)
